+++ transaction/src/validation/transaction_validator.rs
@@ -39,8 +39,8 @@ pub struct ValidationConfig {
     pub network_id: u8,
     pub min_cost_unit_limit: u32,
     pub max_cost_unit_limit: u32,
-    pub min_tip_percentage: u8,
-    pub max_tip_percentage: u8,
+    pub min_tip_percentage: u16,
+    pub max_tip_percentage: u16,
     pub max_epoch_range: u64,
 }
